+++ R/geom_tree.R
@@ -92,7 +92,7 @@ StatTreeHorizontal <- ggproto("StatTreeHorizontal", Stat,
                                 data
                               },
                               compute_panel = function(self, data, scales, params, layout, lineend,
-                                                       continuous = FALSE, rootnode = FALSE) {
+                                                       continuous = FALSE, rootnode = TRUE) {
                                   .fun <- function(data) {
                                       df <- setup_tree_data(data)
                                       x <- df$x
@@ -106,7 +106,8 @@ StatTreeHorizontal <- ggproto("StatTreeHorizontal", Stat,
                                           ## introduce this paramete in v=1.7.4
                                           ## rootnode = TRUE which behave as previous versions.
                                           ## and has advantage of the number of line segments is consistent with tree nodes.
-                                          ## if rootnode = FALSE, the root to itself line segment will be removed (now the new default).
+                                          ## i.e. every node has its own line segment, even for root.
+                                          ## if rootnode = FALSE, the root to itself line segment will be removed.
 
                                           df <- dplyr::filter(df, .data$node != tidytree:::rootnode.tbl_tree(df)$node)
                                       }
@@ -138,7 +139,7 @@ StatTreeVertical <- ggproto("StatTreeVertical", Stat,
                                 data
                             },
                             compute_panel = function(self, data, scales, params, layout, lineend,
-                                                     continuous = FALSE, rootnode = FALSE) {
+                                                     continuous = FALSE, rootnode = TRUE) {
                                 .fun <- function(data) {
                                     df <- setup_tree_data(data)
                                     x <- df$x
@@ -177,7 +178,7 @@ StatTree <- ggproto("StatTree", Stat,
                         data
                     },
                     compute_panel = function(self, data, scales, params, layout, lineend,
-                                             continuous =  FALSE, rootnode = FALSE) {
+                                             continuous =  FALSE, rootnode = TRUE) {
                         .fun <- function(data) {
                             df <- setup_tree_data(data)
                             x <- df$x